% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{cutoff.choose}
\alias{cutoff.choose}
\title{Choose the cutoff corresponding to desired}
\usage{
cutoff.choose(roc, col.nm, value)
}
\arguments{
\item{roc}{roc table, list}

\item{col.nm}{column name of the roc table, string}

\item{value}{value of desired}
}
\value{
list with cutoff value and line of the roc table for this cutoff
}
\description{
Choose the cutoff corresponding to desired
}
