% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{dat.quantile.transf}
\alias{dat.quantile.transf}
\title{Normalize data using quantiles}
\usage{
dat.quantile.transf(S, Y)
}
\arguments{
\item{S}{surrogate S}

\item{Y}{labels containing NA}
}
\value{
dataframe(S.new,Y) where S.new is the quantile transform of S
}
\description{
Normalize data using quantiles
}
