% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{ini.FUN}
\alias{ini.FUN}
\title{computes initial values in EM algorithm}
\usage{
ini.FUN(S, Y)
}
\arguments{
\item{S}{surrogate S}

\item{Y}{labels containing NA}
}
\value{
probability estimates alpha1, alpha0
}
\description{
computes initial values in EM algorithm
}
