% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_methods.R
\name{roc.semi.superv}
\alias{roc.semi.superv}
\title{Transform data and apply EM algorithm, return the auc and full ROC table}
\usage{
roc.semi.superv(S, Y)
}
\arguments{
\item{S}{surrogate S}

\item{Y}{labels containing NA}
}
\value{
list with auc and roc table
}
\description{
Transform data and apply EM algorithm, return the auc and full ROC table
}
