% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{roc.superv}
\alias{roc.superv}
\title{Apply a classic supervised method to return the AUC and full ROC table}
\usage{
roc.superv(S, Y)
}
\arguments{
\item{S}{surrogate of the small training set (all labeled)}

\item{Y}{labels of the small labeled set (no NA)}
}
\value{
list with auc and roc table
}
\description{
Apply a classic supervised method to return the AUC and full ROC table
}
