% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{roc.semi.superv}
\alias{roc.semi.superv}
\title{Transform the data using quantiles and apply EM algorithm, return the auc and full ROC table (semi-supervised learning)}
\usage{
roc.semi.superv(S, Y)
}
\arguments{
\item{S}{surrogate S}

\item{Y}{labels containing NA}
}
\value{
list with auc and roc table
}
\description{
Transform the data using quantiles and apply EM algorithm, return the auc and full ROC table (semi-supervised learning)
}
